+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - Added repository codeowners
 - Added tests for queries with datasetAlleleResponses == HIT and MISS
 - No conflicts between queried assembly and the assembly or queried datasets
+- Parse SVs and save them to database